refactor: split snapshot helper modules (utils / close intent / text runtime)
